route-optimizer/requirements.txt
# Route Optimizer - Python OR-Tools Microservice
# Dependencies

# OR-Tools - Google's Operations Research library
ortools>=9.8.3296

# FastAPI - Modern web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Pydantic for data validation
pydantic>=2.5.3

# HTTP client for external APIs (Google Maps, etc.)
httpx>=0.26.0

# Geolocation utilities
geopy>=2.4.1

# Environment variables
python-dotenv>=1.0.0

# Logging
structlog>=24.1.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
